requirements.txt
PyYAML==6.0
SQLAlchemy==1.4.36
requests==2.27.1
numpy==1.21.6
pandas==1.3.5
Flask==2.1.2
Flask-Caching==1.10.1
Flask-SQLAlchemy==2.5.1
Jinja2==3.1.2
geopy==2.2.0
records==0.5.3
click==8.1.3
flask_googlemaps==0.2.4
detect_delimiter==0.1.1
jinja2schema==0.1.4
plotly==5.8.0